pluto: rename and fixup initiate_replace() to handle_expiring_sa()
If the Child SA is not the newest_ipsec_sa, don't try to initiate
a new one, but just delete it.

Note a comment removed in the code said:

	Fix ME EVENT_SA_REKEY is the not the right one it set next timer 8 hours

This is correct though, as the default IKE and IPsec SA lifetime is now 8h
since:

commit 6b7b51d6cefb77d71a4fd639c9ef6aa2e3145dc6
Author: Paul Wouters <pwouters@redhat.com>
Date:   Fri Jan 22 11:17:40 2021 -0500

    pluto: change default IKE SA lifetime from 1h to 8h.

While it looked to me that EVENT_SA_EXPIRE would be the more natural
event for this, it has as side effect that it first deletes, before it
initiates a replacement, and thus causes a traffic interruption and a
ACQUIRE from the kernel because we have a %trap policy.

logging was added to explain what was done with the EXPIRE message.