pluto: display actual ESPmax= and AHmax= values
This was setting values from st->st_esp.attrs.life_kilobytes but
those are never set. Instead, pull them from c->sa_ipsec_life_bytes

Note that it currently displays ESPmax=17592186044415MB for unset.
While we can fix this, we will soon we overriding this (int64_t) -1
with the FIPS / IETF maximums for 3DES (2^16) and AES (2^32 ?)

Note that this is another problem. We might not know if the connection
will come up with 3DES or AES, so we can only fill in those
maximums later in struct kernel_sa, but then we cannot use
c->sa_ipsec_life_bytes as a reference either.